+++ pg-core/src/java/org/pg/Main.java
@@ -20,19 +20,21 @@ public final class Main {
 //                .build();
 
         ConnConfig config = ConnConfig.builder("test", "test")
-                .port(10160)
+                .port(10140)
                 .host("127.0.0.1")
                 .password("test")
-                .binaryEncode(true)
-                .binaryDecode(true)
+                .binaryEncode(false)
+                .binaryDecode(false)
                 .SOKeepAlive(true)
                 .build();
 
         // Connection conn = new Connection("127.0.0.1", 15432, user, user, user);
         Connection conn = new Connection(config);
 
-        System.out.println(conn.getId());
-        System.out.println(conn.getPid());
+        //System.out.println(conn.getId());
+        //System.out.println(conn.getPid());
+
+        // System.out.println(conn.execute("select '{\"foo\": 555}'::jsonb as obj"));
 
         // System.out.println(conn.execute("select '1 year 1 second'::interval as interval"));
 
@@ -66,13 +68,15 @@ public final class Main {
 
         // System.out.println(conn.execute(""));
 
+        conn.execute("insert into test_json (data) values ($1)", ExecuteParams.builder().params(List.of("[1, 2, 3]")).build());
+
         String query = "select $1 as foo";
         // "select $1::int as foo, 'test' as string, 42 as num, now() as now"
         PreparedStatement ps = conn.prepare(query, ExecuteParams.builder().OIDs(List.of(OID.BOOL)).build());
         // List<Object> params = List.of(1);
         Object res2 = conn.executeStatement(ps, ExecuteParams.builder().params(List.of(true)).build());
         conn.closeStatement(ps);
-        System.out.println(res2.toString());
+        // System.out.println(res2.toString());
 
           //Object res3 = conn.execute("select 'ёёёё'::char as char");
           //System.out.println(res3);