+++ pg-core/src/java/org/pg/clojure/LazyMap.java
@@ -91,8 +91,8 @@ public final class LazyMap extends APersistentMap {
 
         final Object value = switch (col.format()) {
             case TXT -> {
-                final String string = new String(payload, offset, length, codecParams.serverCharset);
-                yield DecoderTxt.decode(string, col.typeOid());
+                final String string = new String(payload, offset, length, codecParams.serverCharset());
+                yield DecoderTxt.decode(string, col.typeOid(), codecParams);
             }
             case BIN -> {
                 final ByteBuffer buf = ByteBuffer.wrap(payload, offset, length);